+++ feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/mapper/response/AfternoteListResponseToListPage.kt
@@ -1,6 +1,6 @@
 package com.afternote.feature.afternote.data.mapper.response
 
-import com.afternote.feature.afternote.data.dto.response.AfternoteListResponse
+import com.afternote.feature.afternote.data.dto.AfternoteListResponse
 import com.afternote.feature.afternote.data.mapper.toDomainList
 import com.afternote.feature.afternote.domain.model.ListPage
