+++ feature/afternote/presentation/src/test/java/com/afternote/presentation/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package com.afternote.presentation
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
